refactor: move factories to its own entities folder
