+++ src/main/factories/controllers/account/signup/signup-controller-factory.ts
@@ -1,7 +1,7 @@
 import { SignUpController } from '@src/presentation/controllers/account/signup/signup-controller'
 import { makeSignUpValidations } from './signup-validations-factory'
 import { makeDbAddAccountUseCase } from '@src/main/factories/usecases/authentication/db-authentication-usecase-factory'
-import { makeDbAuthenticationUseCase } from '@src/main/factories/usecases/add-account/db-add-account-usescase-factory'
+import { makeDbAuthenticationUseCase } from '@src/main/factories/usecases/account/add-account/db-add-account-usescase-factory'
 import { makeLogControllerDecorator } from '@src/main/factories/controllers/decorators/log-controller-decorator-factory'
 
 export const makeSignUpController = () => {
