+++ src/main/factories/controllers/middlewares/auth/auth-middleware-factory.ts
@@ -1,4 +1,4 @@
-import { makeDbLoadAccountByTokenUsecase } from '@src/main/factories/usecases/load-account-by-token/db-load-account-by-token-factory'
+import { makeDbLoadAccountByTokenUsecase } from '@src/main/factories/usecases/account/load-account-by-token/db-load-account-by-token-factory'
 import { AuthMiddleware } from '@src/presentation/middlewares/auth-middleware'
 import { Controller } from '@src/presentation/protocols'
